Restore dict-structures.h to what it used to be
Its an extrnal API file; it can't be garbaged up.
