+++ link-grammar/dict-common/regex-morph.c
@@ -18,6 +18,7 @@
 #include "error.h"          /* verbosity */
 #include "externs.h"        /* lgdebug() */
 #include "dict-api.h"
+#include "dict-common.h"
 #include "link-includes.h"
 #include "regex-morph.h"
