+++ link-grammar/dict-file/read-regex.c
@@ -11,7 +11,7 @@
 #include <string.h>
 #include "link-includes.h"
 #include "api-structures.h"
-#include "dict-common/dict-structures.h"
+#include "dict-common/dict-common.h"
 #include "dict-common/file-utils.h"
 #include "read-regex.h"
